change: remove modules on remove bot

+++ src/application/Guild.ts
@@ -4,6 +4,7 @@ import GuildRepository from "../domain/guild/GuildRepository.mongo";
 import ScheduleUnmuteRepository from "../domain/schedule/unmute/ScheduleUnmuteRepository.mongo";
 
 import type { LocaleLabel } from "../locale";
+import WelcomeModuleRepository from "../domain/modules/welcomeModule/WelcomeModuleRepository.mongo";
 
 class Guild {
     static async create(guildId: string, locale: LocaleLabel = 'en-US') {
@@ -24,10 +25,15 @@ class Guild {
     static async delete(guildId: string) {
         const session = await mongoose.startSession();
         try {
-            session.startTransaction();
-            await ScheduleUnmuteRepository.deleteManyByGuildId(guildId, session);
-            await GuildRepository.delete(guildId, session);
-            await session.commitTransaction();
+            const guild = await GuildRepository.findById(guildId);
+            if (guild) {
+                session.startTransaction();
+                await ScheduleUnmuteRepository.deleteManyByGuildId(guild._id, session);
+                await GuildRepository.delete(guildId, session);
+                if (typeof guild.modules?.welcomeMember?.settings === 'string')
+                    await WelcomeModuleRepository.delete(guild.modules?.welcomeMember?.settings, { session });
+                await session.commitTransaction();
+            }
         } catch (error) {
             await session.abortTransaction();
             throw error;
